+++ code/modules/admin/verbs/mapping.dm
@@ -20,16 +20,25 @@
 //- Check if the area has too much empty space. If so, make it smaller and replace the rest with maintenance tunnels.
 
 GLOBAL_LIST_INIT(admin_verbs_debug_mapping, list(
+	/client/proc/camera_view, 				//-errorage
+	/client/proc/sec_camera_report, 		//-errorage
 	/client/proc/intercom_view, 			//-errorage
+	/client/proc/air_status, //Air things
+	/client/proc/Cell, //More air things
+	/client/proc/atmosscan, //check plumbing
+	/client/proc/powerdebug, //check power
 	/client/proc/count_objects_on_z_level,
 	/client/proc/count_objects_all,
 	/client/proc/cmd_assume_direct_control,	//-errorage
+	/client/proc/startSinglo,
 	/client/proc/set_server_fps,	//allows you to set the ticklag.
+	/client/proc/cmd_admin_grantfullaccess,
 	/client/proc/cmd_admin_areatest_all,
 	/client/proc/cmd_admin_areatest_station,
 	#ifdef TESTING
 	/client/proc/see_dirty_varedits,
 	#endif
+	/client/proc/cmd_admin_test_atmos_controllers,
 	/client/proc/cmd_admin_rejuvenate,
 	/datum/admins/proc/show_traitor_panel,
 	/client/proc/disable_communication,
@@ -40,7 +49,8 @@ GLOBAL_LIST_INIT(admin_verbs_debug_mapping, list(
 	/client/proc/stop_line_profiling,
 	/client/proc/show_line_profiling,
 	/client/proc/create_mapping_job_icons,
-	/client/proc/debug_z_levels
+	/client/proc/debug_z_levels,
+	/client/proc/place_ruin
 ))
 GLOBAL_PROTECT(admin_verbs_debug_mapping)
 
@@ -48,7 +58,7 @@ GLOBAL_PROTECT(admin_verbs_debug_mapping)
 	name = "map fix marker"
 	icon = 'icons/mob/screen_gen.dmi'
 	icon_state = "mapfixmarker"
-	desc = ""
+	desc = "I am a mappers mistake."
 
 /obj/effect/debugging/marker
 	icon = 'icons/turf/areas.dmi'
@@ -57,8 +67,29 @@ GLOBAL_PROTECT(admin_verbs_debug_mapping)
 /obj/effect/debugging/marker/Move()
 	return 0
 
+/client/proc/camera_view()
+	set category = "Mapping"
+	set name = "Camera Range Display"
+
+	var/on = FALSE
+	for(var/turf/T in world)
+		if(T.maptext)
+			on = TRUE
+		T.maptext = null
+
+	if(!on)
+		var/list/seen = list()
+		for(var/obj/machinery/camera/C in GLOB.cameranet.cameras)
+			for(var/turf/T in C.can_see())
+				seen[T]++
+		for(var/turf/T in seen)
+			T.maptext = "[seen[T]]"
+	SSblackbox.record_feedback("tally", "admin_verb", 1, "Show Camera Range") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
+	SSblackbox.record_feedback("tally", "admin_verb", 1, "Show Camera Range")
+
 #ifdef TESTING
 GLOBAL_LIST_EMPTY(dirty_vars)
+
 
 /client/proc/see_dirty_varedits()
 	set category = "Mapping"
@@ -74,6 +105,46 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 	popup.open()
 #endif
 
+/client/proc/sec_camera_report()
+	set category = "Mapping"
+	set name = "Camera Report"
+
+	if(!Master)
+		alert(usr,"Master_controller not found.","Sec Camera Report")
+		return 0
+
+	var/list/obj/machinery/camera/CL = list()
+
+	for(var/obj/machinery/camera/C in GLOB.cameranet.cameras)
+		CL += C
+
+	var/output = {"<B>Camera Abnormalities Report</B><HR>
+<B>The following abnormalities have been detected. The ones in red need immediate attention: Some of those in black may be intentional.</B><BR><ul>"}
+
+	for(var/obj/machinery/camera/C1 in CL)
+		for(var/obj/machinery/camera/C2 in CL)
+			if(C1 != C2)
+				if(C1.c_tag == C2.c_tag)
+					output += "<li><font color='red'>c_tag match for cameras at [ADMIN_VERBOSEJMP(C1)] and [ADMIN_VERBOSEJMP(C2)] - c_tag is [C1.c_tag]</font></li>"
+				if(C1.loc == C2.loc && C1.dir == C2.dir && C1.pixel_x == C2.pixel_x && C1.pixel_y == C2.pixel_y)
+					output += "<li><font color='red'>FULLY overlapping cameras at [ADMIN_VERBOSEJMP(C1)] Networks: [json_encode(C1.network)] and [json_encode(C2.network)]</font></li>"
+				if(C1.loc == C2.loc)
+					output += "<li>Overlapping cameras at [ADMIN_VERBOSEJMP(C1)] Networks: [json_encode(C1.network)] and [json_encode(C2.network)]</li>"
+		var/turf/T = get_step(C1,turn(C1.dir,180))
+		if(!T || !isturf(T) || !T.density )
+			if(!(locate(/obj/structure/grille) in T))
+				var/window_check = 0
+				for(var/obj/structure/window/W in T)
+					if(W.dir == turn(C1.dir,180) || (W.dir in list(5,6,9,10)))
+						window_check = 1
+						break
+				if(!window_check)
+					output += "<li><font color='red'>Camera not connected to wall at [ADMIN_VERBOSEJMP(C1)] Network: [json_encode(C1.network)]</font></li>"
+
+	output += "</ul>"
+	usr << browse(output,"window=airreport;size=1000x500")
+	SSblackbox.record_feedback("tally", "admin_verb", 1, "Show Camera Report") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
+
 /client/proc/intercom_view()
 	set category = "Mapping"
 	set name = "Intercom Range Display"
@@ -84,15 +155,21 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 	for(var/obj/effect/debugging/marker/M in world)
 		qdel(M)
 
+	if(intercom_range_display_status)
+		for(var/obj/item/radio/intercom/I in world)
+			for(var/turf/T in orange(7,I))
+				var/obj/effect/debugging/marker/F = new/obj/effect/debugging/marker(T)
+				if (!(F in view(7,I.loc)))
+					qdel(F)
 	SSblackbox.record_feedback("tally", "admin_verb", 1, "Show Intercom Range") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
 
 /client/proc/cmd_show_at_list()
 	set category = "Mapping"
 	set name = "Show roundstart AT list"
-	set desc = ""
+	set desc = "Displays a list of active turfs coordinates at roundstart"
 
 	var/dat = {"<b>Coordinate list of Active Turfs at Roundstart</b>
-	 <br>Real-time Active Turfs list you can see in Air Subsystem at active_turfs var<br>"}
+		<br>Real-time Active Turfs list you can see in Air Subsystem at active_turfs var<br>"}
 
 	for(var/t in GLOB.active_turfs_startlist)
 		var/turf/T = t
@@ -106,7 +183,7 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 /client/proc/cmd_show_at_markers()
 	set category = "Mapping"
 	set name = "Show roundstart AT markers"
-	set desc = ""
+	set desc = "Places a marker on all active-at-roundstart turfs"
 
 	var/count = 0
 	for(var/obj/effect/abstract/marker/at/AT in GLOB.all_abstract_markers)
@@ -128,21 +205,21 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 	set name = "Debug verbs - Enable"
 	if(!check_rights(R_DEBUG))
 		return
-	verbs -= /client/proc/enable_debug_verbs
-	verbs.Add(/client/proc/disable_debug_verbs, GLOB.admin_verbs_debug_mapping)
+	remove_verb(src, /client/proc/enable_debug_verbs)
+	add_verb(src, list(/client/proc/disable_debug_verbs, GLOB.admin_verbs_debug_mapping))
 	SSblackbox.record_feedback("tally", "admin_verb", 1, "Enable Debug Verbs") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
 
 /client/proc/disable_debug_verbs()
 	set category = "Debug"
 	set name = "Debug verbs - Disable"
-	verbs.Remove(/client/proc/disable_debug_verbs, GLOB.admin_verbs_debug_mapping)
-	verbs += /client/proc/enable_debug_verbs
+	remove_verb(src, list(/client/proc/disable_debug_verbs, GLOB.admin_verbs_debug_mapping))
+	add_verb(src, /client/proc/enable_debug_verbs)
 	SSblackbox.record_feedback("tally", "admin_verb", 1, "Disable Debug Verbs") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
 
 /client/proc/count_objects_on_z_level()
 	set category = "Mapping"
 	set name = "Count Objects On Level"
-	var/level = input("Which z-level?","Level?") as text|null
+	var/level = input("Which z-level?","Level?") as text
 	if(!level)
 		return
 	var/num_level = text2num(level)
@@ -151,7 +228,7 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 	if(!isnum(num_level))
 		return
 
-	var/type_text = input("Which type path?","Path?") as text|null
+	var/type_text = input("Which type path?","Path?") as text
 	if(!type_text)
 		return
 	var/type_path = text2path(type_text)
@@ -182,7 +259,7 @@ GLOBAL_LIST_EMPTY(dirty_vars)
 	set category = "Mapping"
 	set name = "Count Objects All"
 
-	var/type_text = input("Which type path?","") as text|null
+	var/type_text = input("Which type path?","") as text
 	if(!type_text)
 		return
 	var/type_path = text2path(type_text)
@@ -220,12 +297,19 @@ GLOBAL_VAR_INIT(say_disabled, FALSE)
 	D.setDir(SOUTH)
 	for(var/job in subtypesof(/datum/job))
 		var/datum/job/JB = new job
-		for(var/obj/item/I in D)
-			qdel(I)
-		randomize_human(D)
-		JB.equip(D, TRUE, FALSE)
-		var/icon/I = icon(getFlatIcon(D), frame = 1)
-		final.Insert(I, JB.title)
+		switch(JB.title)
+			if("AI")
+				final.Insert(icon('icons/mob/ai.dmi', "ai", SOUTH, 1), "AI")
+			if("Cyborg")
+				final.Insert(icon('icons/mob/robots.dmi', "robot", SOUTH, 1), "Cyborg")
+			else
+				for(var/obj/item/I in D)
+					qdel(I)
+				randomize_human(D)
+				JB.equip(D, TRUE, FALSE)
+				COMPILE_OVERLAYS(D)
+				var/icon/I = icon(getFlatIcon(D), frame = 1)
+				final.Insert(I, JB.title)
 	qdel(D)
 	//Also add the x
 	for(var/x_number in 1 to 4)
